rename configs and packages
