+++ src/main/resources/reference.conf
@@ -4,24 +4,22 @@
 # Apache Pekko Service Discovery Consul Config       #
 ######################################################
 
-pekko.discovery {
+pekko-discovery-consul {
 
   # Set the following in your application.conf if you want to use this discovery mechanism:
   # impl = pekko-consul
   # ApiMayChange
-  pekko-consul {
-    class = org.apache.pekko.discovery.consul.ConsulServiceDiscovery
+  class = com.github.pjfanning.pekko.discovery.consul.ConsulServiceDiscovery
 
-    consul-host = "127.0.0.1"
-    consul-port = 8500
+  consul-host = "127.0.0.1"
+  consul-port = 8500
 
-    # Prefix for consul tag with the name of the actor system / application name,
-    # services with this tag present will be found by the discovery mechanism
-    # i.e. `system:test` will be found in cluster if the cluster system is named `test`
-    application-name-tag-prefix = "system:"
+  # Prefix for consul tag with the name of the actor system / application name,
+  # services with this tag present will be found by the discovery mechanism
+  # i.e. `system:test` will be found in cluster if the cluster system is named `test`
+  application-name-tag-prefix = "system:"
 
-    # Prefix for tag containing port number where Pekko management is set up so that
-    # the seed nodes can be found, an example value for the tag would be `pekko-management-port:19999`
-    application-pekko-management-port-tag-prefix = "pekko-management-port:"
-  }
+  # Prefix for tag containing port number where Pekko management is set up so that
+  # the seed nodes can be found, an example value for the tag would be `pekko-management-port:19999`
+  application-pekko-management-port-tag-prefix = "pekko-management-port:"
 }
